Rename SliceInterface.ts -> Slices.ts
