+++ api/services/apdf/src/providers/APDFRepositoryProvider.ts
@@ -11,7 +11,7 @@ import {
 import { APDFTripInterface } from '../interfaces/APDFTripInterface';
 import { PolicyStatsInterface } from '../shared/apdf/interfaces/PolicySliceStatInterface';
 import { PgCursorHandler } from '../shared/common/PromisifiedPgCursor';
-import { UnboundedSlices } from '../shared/policy/common/interfaces/SliceInterface';
+import { UnboundedSlices } from '../shared/policy/common/interfaces/Slices';
 
 @provider({ identifier: DataRepositoryProviderInterfaceResolver })
 export class DataRepositoryProvider implements DataRepositoryInterface {
