+++ api/services/apdf/src/providers/excel/wrapSlicesHelper.ts
@@ -1,4 +1,4 @@
-import { BoundedSlices, UnboundedSlices } from '../../shared/policy/common/interfaces/SliceInterface';
+import { BoundedSlices, UnboundedSlices } from '../../shared/policy/common/interfaces/Slices';
 
 export function wrapSlices(slices: BoundedSlices | UnboundedSlices | undefined | null): UnboundedSlices | [] {
   if (!slices || !Array.isArray(slices) || !slices.length) return [];
